Fix integer overflow lint errors (#4227)
Fixed through a mixture of type changes, and ignoring the linter in
places where I was confident that we would never hit an overflow in
practice.

+++ internal/history/service.go
@@ -204,7 +204,7 @@ func (_ *evaluationHistoryService) ListEvaluationHistory(
 	filter ListEvaluationFilter,
 ) (*ListEvaluationHistoryResult, error) {
 	params := db.ListEvaluationHistoryParams{
-		Size: int32(size),
+		Size: int64(size),
 	}
 
 	if err := toSQLCursor(cursor, &params); err != nil {
